Changed project layout Changed tests to illustrate another way of laying them out.
